[Bug] Reads do not return credentials, and lead to replacement
Ambar describe API's will omit credentials to protect users. Given credentials
are passed as part of a DataSourceConfiguration, this can lead to Terraform
detecting the omission and triggering a replacement of the resource. Instead
we should carry the credentials from state forward, and then in further plan
steps Terraform can decide if the user changed the creds and the resource needs
replacement.