Require replace on FAILED state
When a DataSource is in the FAILED state, if there is any other change to the
resource to trigger an apply graph check, we will force a resoruce replacement.
DataSources anyways currently require replacement on all attributes at the moment
but when credential rotation or other updates are supported this will still be
needed.